SSA cleanup.
